+++ src/tables/cmap/format13.rs
@@ -4,7 +4,9 @@ use core::convert::TryFrom;
 
 use crate::parser::Stream;
 
-pub fn parse(mut s: Stream, code_point: u32) -> Option<u16> {
+pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
+    let mut s = Stream::new(data);
+    s.skip::<u16>(); // format
     s.skip::<u16>(); // reserved
     s.skip::<u32>(); // length
     s.skip::<u32>(); // language
@@ -18,4 +20,10 @@ pub fn parse(mut s: Stream, code_point: u32) -> Option<u16> {
     }
 
     None
+}
+
+pub fn codepoints(data: &[u8], f: impl FnMut(u32)) -> Option<()> {
+    // Only the glyph id mapping differs for this table. The code points are the
+    // same as for format 12.
+    super::format12::codepoints(data, f)
 }